+++ semcore/inline-input/CHANGELOG.md
@@ -2,12 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [2.0.1] - unreleased
+## [2.1.0] - unreleased
 
 ### Changed
 
 - Synced colors with design.
 - Moved all color definitions to themable styles.
+
+## [2.0.1] - 2022-04-28
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/icon` [2.20.0 ~> 2.21.0]).
 
 ## [2.0.0] - 2022-04-07